Added empty unittest framework; moved summary to top of outfile.
Added an empty unittest framework (still need to implement all tests),
with a -t option to trigger tests and appropriate code to run them.

Moved the summary to the top of the output file (probably makes more
sense);  changed the order of printing things a bit to make sure all
combinations of -H and -s/-S arguments look reasonable.

Fixed bug in the verbose messages to stdout - filename and timestamp
were switched.