Renamed start/end position types to eliminate confusion; minor changes.
Renamed start/end position types to leftmost/rightmost to eliminate
confusion (even I was confused, since I made "end" the default while I
actually meant "3prime").  Changed the -p option default to "3prime".

Added more information on what these mean to the option help message,
and modified the docstring to mention that the relevant position feature
can be changed depending on options.

Moved the list of position types into a global variable
VALID_POSITION_TYPES, since it was repeated several times in the code,
which is bad; changed most of those parts of the code to use that
variable.  We still have a few places where '3prime' etc are repeated,
but I think that's inevitable (the actual code that defines position
based on pos_type, explanations of the types, default option value).

Added an important TODO: we DON'T want the grouping to be based only on
alignment location, I think!  What if we have two mutants that inserted
into the same location but in opposite orientations?  12--->345 and
123<---45 - position should be "3" in both cases (is that really how it
comes out? check!), but I think those are separate mutants and should be
treated separately, right?  Unlikely, of course, but still.