Added option for getting correct orientation from reverse reads.
Added an option for when reads are in reverse orientation to the
cassette, to make sure the cassette-vs-gene sense/antisense orientation
is still correct (if read is reverse, then if read and gene are on the
same strand, the orientation is actually antisense!)

In deepseq_analysis_classes.py, added a read_is_reverse argument to
find_gene_by_pos and implemented switching the results (by switching a
temporary copy of the gene strand value) if the argument is True;
rewrote All_alignments_grouped_by_pos __init__ to take read_is_reverse
as an argument and save it (also added some organization/comments
there), and made add_gene_positions_to_data pass that value on to
find_gene_by_pos.
In deepseq_count_alignments.py, added an option for reverse reads
(-r/-R), and passed it on to the All_alignments_grouped_by_pos
constructor call.

TODO should probably write a test-run case for this.