Added an output line giving the largest group size as a sanity check.
Added a line to the output that gives the largest group position, read
number, and percentage of reads compared to total aligned reads, as a
sanity check - sometimes there's a single group responsible for a large
number of the reads, which is an important thing to know, as it can
make the various counts in the summary look strange.

Examples of the new output line:
 - sample L2: "Most common group: chromosome_12, position 5268560, +
   strand: 19800 reads (5% of aligned)" - 5% isn't too bad
 - sample L7: "Most common group: chromosome_12, position 5436099, +
   strand: 1806293 reads (44% of aligned)" - 44% is bad! Explains the
   unequal +/- strand totals, etc.

Currently the run-tests don't pass because I haven't added that line to
the test output reference files yet - I'll make some more output changes
and then make the modifications all at once.