Changed --read_cassette_end default to '5prime' (the correct value).
The MmeI site we're using right now is on the 5prime end of the cassette
(i.e. the "front"), so this should be the default.

Note that there's currently an issue with the way I determine cassette
position when the read aligns to the - strand - see TODO in
Insertion_position __init__!