Fixed the problem with - strand reads getting wrong positions.
See the "Possible read sides/directions" section in ../notes.txt for how
cassette position depends on read position - it's rather complicated!

I realized there was a problem when I saw the real-world L2 data
analysis after the rewrite reported over a hundred more insertions than
the old analysis.  I looked at them and realized that whenever the
cassette was on the - strand and had multiple different-length read
sequences (20 and 21bp, due to MmeI cut position imprecision), it would
be reported as two insertions 1bp apart from each other.  Clearly in
that case the wrong end of the read was being used as the cassette
position.  The problem was that I thought the read orientation didn't
matter, but actually it's just the read orientation WITH RESPECT TO THE
CASSETTE didn't matter, but the CASSETTE orientation (+ or - strand) DID
matter: if the read is 5' of the cassette and the cassette inserted
forward (+ strand), the position of the cassette is read.end-?, but if
the read is 5' and the cassette inserted reverse (- strand), the
position of the cassette is ?-read.start.

I rewrote all the run-test cases to reflect that (although I still got
the reverse output file wrong!), then rewrote the code to make them
pass.

The main changes were in get_insertion_pos_from_HTSeq_read_position:
included the cassette (not read) strand in determining the positions (so
instead of "if cassette_end=='5prime': pos_5prime, pos_3prime =
HTSeq_pos.end, None" we have "if (cassette_end=='5prime' and
strand=='+') or (cassette_end=='3prime' and strand=='-'): pos_before,
pos_after = HTSeq_pos.end, None", and the same for the other end/strand
combination.  Rewrote comments/docstring to make it clear how this
worked.

Changed Insertion_position attribute names from
position_5prime/position_3prime to position_before/position_after, which
more accurately reflect what I'm actually reporting here - the two
positions are the same regardless of cassette orientation.

Changed/added unit-tests for get_insertion_pos_from_HTSeq_read_position.

ALL RUN-TESTS AND UNIT-TESTS PASS.

I ran the new version on the real-world L2 data, and the output is
identical to the old version (after removing the full_position column)
except for summary wording differences.  Is that a good thing?...  I
suppose actually it's as expected - when I originally thought that
insertions in opposite directions in the same location would get
assigned positions 1bp off, I didn't realize HTSeq used a 0-based
end-excluded position scheme (i.e. the end position is actually the
position AFTER the last base).  I suppose that's reasonable.  At least
now I know my code is doing the right thing on purpose as opposed to by
accident, and that it'll still work even if the reads are on the other
end of the cassette or in reverse orientation to the cassette, which I'm
not sure would have been the case before.