Added read direction info to summary, minor formatting changes.
Added the read direction (forward/reverse with respect to cassette) info
to summary; moved that and the cassette_end info (which end of the
cassette the read is on) to a separate line.

Changed the header to name the sequence/count fields
read_sequence_X/seq_X_count instead of sequence_X/count_seq_X, to make
it clearer that they're read sequences, not cassette sequences.

Changed all run-test reference output files to match.  All run-tests and
unit-tests pass.