Fixed but in data sorting to sort chr11 after chr2.
There was an error in the regular expression used to split chromosome
into chromosome_name and chromosome_number in Insertion_position that
made it always use the full string as the name and set the number to 0,
so that chromosomes got sorted as strings instead of numbers, thus 11
before 2.

The regular expression was originally "^(.*[^\d])(\d*)", which worked; I
changed it to "^(.*[^\d]?)(\d*)" at some point to make it stop throwing
an error in a unit-test that had only a number as a chromosome name, but
that actually made it not work as desired, because a non-digit character
wasn't required in the first group, so the whole string always matched
the first group (regexes are greedy).
Got around that by using named groups instead of numbered ones, and
making the whole first group (instead of just the non-digit character)
optional - now if there are no non-digit characters, the first group
will be skipped and the second group will match the entire string, which
is what I want.  Then calling groups('name') will return None, since
that group wasn't matched, which is fine - None sorts before strings,
like ''.

All unit-tests and run-tests pass, analysis on real-world L2 data shows
chromosomes sorting correctly.