More complex gene features, incl. edge cases; separate gene edge case.
(Changes in mutant_analysis_classes.py unless otherwise specified)

Rewrote find_gene_by_pos in to deal with gene features in more detail,
including overlaps between multiple features and edges of a feature.
Now the logic is explicitly based on the expected
gene->mRNA->(exons/UTRs) structure of the gff3 file: the cases of the
gene having no features and the insertion just not matching any of them
are separated; the introns are taken into account even though they're
not explicitly listed; etc.  See function docstring/comments for
detailed logic.  Will also raise exceptions or print warnings if it sees
anything unexpected.

Also rewrote find_gene_by_pos to treat an insertion at the edge of a
gene (overlapping but not contained) as a separate case, instead of
taking a location_in_gene_test function as an argument and just using
that (that function would be either a strict/contains or a
liberal/overlaps check).

Since the location_in_gene_test argument is not necessary now, removed
it; likewise removed the liberal_position_test argument from ___ in
Insertional_mutant_library_dataset, and the liberal_position_test
(-x/-X) optparse options from mutant_count_alignments.py.  Now the user
doesn't have to make that decision, which is good, because in reality
the user should probably almost always run it with both options and
compare the results - now the program does it for her, and she can just
ignore the edge-cases in the output if she wants to.

Checked that BCBio.GFF uses 0-based-end-exclusive positions, while I use
1-based-end-inclusive - adjusted all the positions to reflect this.

Changed the orientation/feature return values when no gene was found or
the chromosome wasn't in the reference from '?' to '-', since it's not
that they're unknown, it's that they're known not to exist.

All unit-tests and run-tests pass; real-world L2 data analysis gives the
following feature counts:
    171 -
     75 intron
     66 exon
     48 three_prime_UTR
     16 five_prime_UTR
      1 five_prime_UTR/exon

TODO add feature-type counts to the summary; write a gene-info run-test
case!