Wrote test_reference.gff3 file with a lot of normal/strange/wrong cases.
The previous version was just a part of the actual chlamy annotation
file reference_data/Creinhardtii_169_gene.gff3.

Contents:
 - chromosome_A: normal-structure genes (on the + and - strand).
 - chromosome_B: genes missing features: no features at all, mRNA only
   (same size as gene or smaller), UTR-only, exon-only, no UTR, no
   exons.
 - chromosome_C: genes with strange feature structures: exon outside the
   UTR, multiple UTRs, UTRs not immediately adjacent to exons.
 - chromosome_D: clearly incorrect genes: mRNA larger than gene, exon
   larger than mRNA, exons with no intron in between, overlapping exons,
   overlapping exon/UTR, two mRNAs (overlapping or not), non-mRNA
   primary feature (made-up type), overlapping genes.

Unfortunately I can't put comments in the file, it makes some of the
BCBio GFF examiner functions fail.

Now I can use this file to write a run-test for the
mutant_count_alignments.py -g and -D options, and also to test
gff_examine_file.py and eventually write a run-test for it.