Checking for gene/mRNA/feature start/end position gaps; changed test file.
Added a check for mRNA not the same size as a gene, and for the start of
the first feature and end of the last feature not matching the mRNA
start/end.  Tested on test_data/test_reference.gff3, detects all cases
that should have it;  the real-world annotation file reference_data/
Creinhardtii_169_gene.gff3 has no cases of either of these things
happening.

In the test_data/test_reference.gff3 test gff3 file:
 - added a geneC5_gap_in_mRNA_before_UTR_start test case gene
 - added a geneD9_bad_feature_outside_mRNA test case gene
 - for all the cases with missing features and with touching/overlapping
   features, modified the gene/mRNA start/end positions to match the
   positions of the features, since this is apparently the norm.  The
   only cases for which this shouldn't be true are the ones specifically
   written for that reason.