Minor: fixed dealing with unlikely implied feature edges.
Changed the dealing with implied features (i.e. non-exon sections) in
find_gene_by_pos function in mutant_analysis_classes.py slightly.
Before the change, if the insertion was in the mRNA but NOT in any
features, it was called 'intron' if it was between two existing
features, or 'mRNA_before_exons'/'mRNA_after_exons' if it wasn't.
However, if it was on the EDGE of ONE feature (i.e. implied edge with an
implied feature), it was always annotated as 'feature/intron', even if
what was supposedly an intron should actually be considered
mRNA_before_exons/mRNA_after_exons.
Fixed that now: if we think there's an implied feature involved (i.e.
insertion is inside mRNA but either not in any features or on the edge
of a single feature), first the type of the implied feature is
determined (intron/mRNA_before_exons/mRNA_after_exons), then the
gene_feature output is set to either just that if there was no edge
feature, or to feature/implied if there was a single edge feature.

Added a test case for that in test_data/test_input2__for-genes.sam
(using the test_data/test_reference.gff3 test genes
geneC5_gap_in_mRNA_before_UTR_start and
geneC6_gap_in_mRNA_after_UTR_end, the latter of which by the way I
renamed from geneC6_gap_in_mRNA_after_UTR_start, which was incorrect);
made a new test_data/test_output2__with-genes.txt to contain the right
output (just auto-generated it and checked the diff with the old
version).   All run-tests and unit-tests pass.