Fixed read_from_file function in Insertional_mutant_library_dataset.
Changed the read_from_file function implementation to match the new
format (including position representation by Insertion_position objects
instead of chrom/strand/pos); added a lot of the dataset total count
info. Rewrote the docstring. Updated the unit-test (and added a new
section with a file with gene info) - it passes.

In order to be able to read a full_position string and make it into a
Insertion_position instance, I rewrote the Insertion_position __init__
function so it can take a full_position string instead of
position_before and/or position_after values, and gave it
parse_single_position and parse_full_position functions.  Changed all
the Insertion_position initiations to specify which position argument is
being given.