Changed the reads_are_reverse default to '?'; other related changes.
Changed the default value of the Insertional_mutant_library_dataset
__init__ argument reads_are_reverse from False to '?' to make it clear
when we have a dataset that DOESN'T specify this (as would happen if
we're reading pre-calculated mutant insertion positions from a file
using read_from_file function.  Also changed the default ("unknown")
value for the cassette_end argument from None to '?'.  Changed the
argument checks in __init__ and the __init__ unit-test to reflect that
(also updated the unit-test to check for new total-count variables).

Changed the check for a non-unknown self.cassette_end value in
add_alignment_reader_to_data to reflect that change, and added another
check for self.reads_are_reverse, since you can't convert alignment data
to mutant insertion positions without that information.
Modified the line in the print_summary function that gives the read
direction to deal with all three possible values (using an inline
dictionary).