Fixed finding the main sequence when datasets disagree on it.
Changed the section of run_main_function that deals with joining the
same mutant from different datasets into one line, specifically the part
that finds the main sequence: if the datasets disagree on what the main
sequence is, the one with the most overall counts between mutants is
taken.  The calculation now takes into account the full seq:count
dictionary of each mutant, not just the most common sequence.

TODO that functionality still hasn't been tested properly!  Although it
appears to work.