Added gene annotation info as extra fields in the output file.
Added a -G/--gene_annotation_file option to take a gene annotation file
(formatted like reference_data/Creinhardtii_169_annotation_info.txt,
tab-separated) and look the gene IDs up in it and add whatever
information is in the file to the end of each line in the output file.
Wrote parse_gene_annotation_file function (long but not terribly
complicated) to get the data, and added sections of run_main_function to
print the data to output file.
Had to add a fair amount of special-casing for the
Creinhardtii_169_annotation_info.txt that we're actually using, while
still leaving the program with the ability to parse any other
tab-separated file that starts with gene IDs.

Made sure all three verbosity levels work sensibly.

TODO write a run-test for this!  For now tested it on real-world data,
works (see outputs_various/joint_counts_with-gene-info.txt)