Padded no-gene-data lines with empty fields to get the same field number.
Instead of just adding a single "NO GENE DATA" field to the lines with
no annotation, add the same number of fields as the lines with gene
annotation have (first field is "NO GENE DATA", the rest are empty), to
make sure the file can be read by simple programs that expect a
tab-separated file with the same number of fields in each line (such as
general_utilities.read_tab_separated_file, which I sometimes use to read
the joint mutant file to do correlation/etc plots).

This was actually done Dec 16, I just forgot to commit until now.