IN PROGRESS: added measuring coverage % of gene features.
Desired outcome: in addition to reporting what % of the genome (and of
each chromosome) is covered by genes, report what % of the total gene
length is covered by exons/UTRs (and the remaining % will be introns).

I implemented this functionality in check_gene_coverage, and modified
the two places where it's used to use and print the feature coverage
info as well as the gene info, but it DOESN'T ACTUALLY OUTPUT ANYTHING,
as explained in the check_gene_coverage docstring, because most of the
time in this program I parse gff files with the genes-only limit, to
save RAM (if I want to include features in the parsing, reading the
whole file takes more RAM than is acceptable for me, so I need to split
it by chromosome and it all gets complicated).  So since no features
have been parsed, genes have no features and the feature coverage output
is an empty dictionary.

TODO rewrite this with including features in parsing the file (and
splitting it into chromosome sets to not parse all at once to save RAM,
like done in the "if options.gene_feature_structure_counts" section), so
that it actually gives the output we want!  Update the
reference_data/Creinhardtii_169_gene_analysis.txt file (and the original
file that's symlinked there) to include that info.