Minor changes: bug fix, text position change, comments/TODOs/etc.
Bug fix for min_value argument when passed to _make_sample_label_text in
two spots (switched != to the correct ==).

Changed the position of the correlation text to something that worked
better with current plots - unsure if it's the best change in general,
we'll see.

Various comments and TODOs, a help message change, etc.