IN PROGRESS: making Insertional_mutant_data class multi-dataset, continued.
I got interrupted by needing to work on another project, so this commit
it mostly work done in early March, with a few simple fixes done on Mar
23 to make it all function.  See previous commit for what the purpose of
all this is.

Removed the old versions of Insertional_mutant_data class methods - the
new ones seemed better.

Added a readcount_related_only argument to Insertional_mutant_data
__init__, to make it possible to generate readcount-only "mutants" that
are actually just elements of the by_dataset dictionary of a
multi-dataset mutant (this gets used in convert_to_multi_dataset and in
the multi-dataset variant of __init__ itself).

Slight rewrites to other Insertional_mutant_data methods:
 - wrote the structure but not the content for get_main_sequence
   multi-dataset functionality - left a TODO to implement.  Added an
   import for tne (new) general_utilities.add_dicts_of_ints it should
   use.
 - rewrote convert_to_multi_dataset slightly, to do nothing if the
   mutant is already multi-dataset, and to make self.by_dataset a
   defaultdict of Insertional_mutant_data(readcount_related_only=True)
   objects (same as the new version of __init__)
 - in add_other_mutant_as_dataset, put a wrapper with a sensible error
   message around the asserts that test the non-readcount-related data
   for equality.
 - added/edited some comments/TODOs/docstrings.

All the old unit-tests and run-tests pass - NOTE that there are still no
tests for the new functionality, but at least the old functionality
still seems to work correctly.