Removed mutant_count_alignments.py no-summary/no-header options.
Removed the silly --header_level and --add_summary_to_file/
--dont_add_summary_to_file command-line options - in reality I ALWAYS
want a header and a summary, no point in adding complexity to the code
(the main reason for the option not to print the header was that it used
to break the run-tests before they were regex-based, but now that's
fine). Modified the data-writing section of run_main_function
accordingly.

Modified all the test output reference files in test_data to match (just
added the four "<REGEX>#" header lines on top, and modified the format
of two other header lines a little.

Modified do_test_run to stop adding "-H 1 -s" to everything, added -n3
to single tests where appropriate, now only -q gets added in the
test-case rewriting line.

All run-test pass.