Added sensible custom dataset order to multi_datasets and mutant_join_datasets.py.
Originally multi-datasets (Insertional_mutant_pool_dataset instances
that were multi-datasets, in mutant_analysis_classes.py) were purely
dictionary-based, so when printing data etc, the dataset columns were in
alphabetical order (since dictionaries are unordered, so I just sorted
the keys to get some kind of reliable order).  This wasn't great when
running mutant_join_datasets.py, since the order of the dataset columns
wouldn't be the same as the order in which the dataset infiles were
given on the command-line.  It got even more annoying when plotting the
dataset correlations, since order matters a lot then.

So I gave a self.dataset_order attribute to multi-dataset
Insertional_mutant_pool_dataset instances, and used that as the order
when printing data (defaulting to alphabetical when not provided, and
raising an error if set(self.dataset_order) isn't the same as the actual set
of datasets in the multi-dataset).
In mutant_join_datasets.py, I set multi_dataset.dataset_order to make it
reflect the original order of the command-line infiles.

I added a run-test case for this in mutant_join_datasets.py, with output
reference file test_data/join-datasets__other-order.txt.  It passes.