Changed "special/ignore chromosomes" options to explicit cassette options.
Now that we have multiple cassette "chromosomes" with different names,
it's easier to just have a function that checks for chromosome names
containing "insertion_cassette", rather than use the general method of
providing a list of "special" chromosomes (I could keep doing that, but
I'd have to keep the default updated with the full list of
insertion_cassette chromosome names, and anyway it seems unnecessarily
general, since there AREN'T any other special chromosomes in our
method.)

So changed the -b/--bad_chromosomes_count_only command-line option
(which took a comma-separated list) to True/False
-I/--dont_count_cassette (later inverted to options.count_cassette, but
the negative is there because the default is True), and similarly
-B/--bad_chromosomes_count_and_ignore to -i/--ignore_cassette.

Rewrote the relevant calls in main (renamed from run_main_function for
simplicity) in mutant_count_alignments.py to use the new options.
Rewrote a few functions in mutant_analysis_classes.py to do the same
(not complicated), with the help of the new is_cassette function.
Nothing at all complicated.

TODO or maybe I should go back to a command-line option that gives a
regex for how to recognize a cassette chromosome, or a list of names?
Or a config file or something... If we ever want to publish the
software, that might be good - for now, who cares.

Only started modifying the run-tests: renamed the test cases and the
test reference files but didn't change them to work with the new code
yet, so the count-aln__dont-count-cassette run-test fails (but all the
ones before it pass).
TODO modify "special" chromosome names in the test files to ones
matching the new is_cassette function - tests should pass then.

Other minor changes:
 - renamed run_main_function to just main, simpler and clearer