Minor bugfix: got rid of useless ignore_cassette arg to find_genes_for_mutants.
Really I think the "if ignore_cassette and not is_cassette(chromosome):
print warning" bit was a bug - why would we NOT want to print the
warning if there a non-cassette unknown chromosome just because
ignore_cassette was false?

New version:  "if not is_cassette(chromosome): print warning", so
cassette chromosomes never get warnings (there's really no point,
they'll never show up in annotation files!) and the whole
ignore_cassette argument is removed - I don't think there was a point to
having it.