Refactoring and adding options for gene feature counts in summary.
Added nicer_gene_feature_counts method to Insertional_mutant_pool_dataset
class - it takes care of some of the sorting/etc functionality that used
to be in print_summary, and adds a feature to merge the feature-boundary
cases together instead of printing a long list of them in the summary.

Added merge_boundary_features arg to print_summary.  If it's False, all
current run-tests pass.  Need to rewrite the run-test reference files to
make it pass when True (and maybe add a command-line option for this to
mutant_count_alignments.py and write run-tests for both versions?).