Changed print_summary to use value_and_percentages from general_utilities.
Since I was using the "%s (%.2g, %.2g)"%(value, value/total1, value/total2)
pattern so much, so I wrote a separate function for it,
general_utilities.value_and_percentages (it takes a value and a list of
totals, and returns a string with the value and its percentage of each
total, like this: 1,[2,3] -> "1 (50%, 33%)").  I rewrote
Insertional_mutant_pool_dataset.print_summary to use it.  It's not
really that much shorter, but it's simpler and more readable.

If I make value_and_percentages return fractions by default (so 1,[2,3]
-> "1 (0.5, 0.33)" instead of the percentage version - using a default
True value for the fractions_not_percentages argument), all the
run-tests pass.
If I change the default to percentages, I'll need to change the run-test
reference files to match (percentages are better anyway, I just did
fractions first to avoid too much repetitive code and for simplicity).