Percentages instead of fractions in run-test reference file summaries.
Modified all the run-test reference files to use percentages instead of
fractions of totals in the summary.
Also changed "fraction of" to "% of" in the summary line description
text, both in the run-test files and in print_summary function in
mutant_analysis_classes.py (note that the strings without a % operator
get % for percent, and the ones with a % operator get %% instead -
result is the same, the implementation is just a little ugly).

Also note that there was one case where the approximation actually
changed when I switched from fractions to percentages:
    >>> 39/40
    0.97499999999999998
    >>> "%.2g"%(39/40)
    '0.97'
    >>> "%.2g"%(100*39/40)
    '98'
Not sure why this is, but it doesn't matter much, so I just changed the
file from 0.97 to 98%.

All run-tests pass.