Various refactoring (in progress, but current version works).
The main motivation is to rewrite the Insertional_mutant_pool_dataset
print_summary() method to make it easy to extend to multi-datasets
(which would currently require a huge amount of difficult-to-maintain
code duplication).  That is about half-finished (see new_print_summary
method), and pre-defines a list of (line_description,
line_value_getter_function) tuples to then print line_description and a
tab-separated list of values, each generated by the value-getter
function for each component dataset (or for the entire dataset if it's
single).

In the process I found several small refactorings:
 - changed dataset_order to be a @property with a getter and setter and
   a check that all the datasets present in summary/mutants are included
   (_check_dataset_consistency method); used that in print_data instead
   of the previous multiple lines defining the dataset order.
 - added a mutants_in_dataset that iterates over only the mutants
   present in a particular dataset, for multi-datasets.
 - some minor changes to comments or order of things or such.

The new_print_summary method is not finished and not used yet, but the
other refactorings are already included in the working parts of the
code. All run-tests (for mutant_count_alignments.py and
mutant_join_datasets.py, which both use mutant_analysis_classes.py)
and unit-tests pass.