ADDED PICKLING - mutant_count_alignments.py makes an extra *.pickle file.
The mutant/dataset objects are getting complicated enough that trying to
write all the relevant data in a file and then parse that file later is
getting ridiculous - I'm going to use the pickle module instead.  So now
each mutant dataset will generate a human-readable *.txt file like
before, AND a *.pickle file that can be read by python to give the full
dataset including all the information that's missing in the text file
(like an unlimited number of different sequences per mutant).

In mutant_analysis_classes.py:
 - In order for that to work, had to remove several lambdas (just
   changed them to named functions, trivial).
 - Added a unit-test for pickling (test__pickle_unpickle; uses the same
   infiles as test__read_data_from_file, with the actual detail checks
   moved into _check_infile1 and _check_infile2 so they can be used by
   both).  PASSES.
 - updated read_data_from_file docstring to explain that it's
   deprecated; removed TODO comments from that function, since it won't
   be updated further.

In mutant_count_alignments.py:
 - added writing a pickle outfile (same filename as normal outfile but
   with .pickle instead of .txt extension) - simple.
 - removed the -O/--mutant_merging_outfile command-line option and made
   the *_merging-info.txt filename standard - no reason not to.
 - renamed the mutant_merging_outfile open file-handle from OUTFILE to
   MERGEFILE for clarity.

Added *.pickle files for all the mutant_count_alignments.py run-test
cases in test_data - they're all symlinks to one file just containing
"<STOP COMPARING FILE>", which is a marker causing that file to compare
True with everything - we don't want to actually check the contents of
the pickle files, that would be ridiculous (could just skip them in the
test, but 1) that would be more code, 2) we do want to make sure they
exist, and 3) explicit is better than implicit).