Refactoring mutant_positions_and_data - single dataset list, 3 formats.
Modified mutant_positions_and_data to take a single list of
different-format datasets, with a dataset_formats list to say which ones
they are, instead of having two separate mutant_datasets and
other_datasets lists and corresponding lists of all other variables.
In addition to the old mutant dataset and chrom:pos_list dataset
formats (0 and 1), added a new chrom:bin_density_list format (2, as
generated by get_histogram_data_from_positions), which will take up much
less memory for huge datasets like mappability data!
This was a fair amout of refactoring, but the basic functionality should
be the same, except for the addition of an extra dataset format, and a
more flexible order of datasets on the plot.
Tested on real data, works!

Other minor things:
 - moved get_chromosome_lengths function up and added better error
   message
 - modified get_histogram_data_from_positions to use
   get_chromosome_lengths to allow chromosome_lengths to be a file, and
   to fail if chromosome lengths are missing