Added colorbar scaling for 0-100% data in mutant_positions_and_data.
Added a maxes_per_base argument to give the number of positions per base
that constitutes 100% (if None, don't do 0-100% scaling, just plot the
raw numbers scaled to whatever the maximum is).  Changed the max_count
value to the correct scaling of that, if given - that's used as max in
the mplt.imshow call.
Modified the colorbars to change their ticks to be 0-100% in those cases.