Added p-values to correlations; other minor.
In addition to reporting correlations, also report the p-values showing
statistical significance (how likely they are to be non-random).

Also refactored some of the parts that make the correlation coefficient
NaN if it's an empty list or something else meaningless.

Also switched from using the numpy Pearson's correlation to the
scipy.stats.stats one.