This was started a few weeks ago, I forgot to commit then.
Three major changes here:

1) Added UTR introns as a separate category from CDS introns - I didn't
previously realize UTR introns existed.

2) Allowed insertions in multiple genes - the v4.3 chlamy genome version
had no overlapping genes, so I didn't have to deal with this, but the
v5.3 one does, so I had to implement it.  The gene/orientation/feature
output fields are |-separated, with info for multiple genes.

3) If gene has multiple splice variants (again, didn't happen in v4.3
chlamy genome, but does in v5.3), there's a warning and the feature is
given as "MULTIPLE_SPLICE_VARIANTS" - TODO later I should make it parse
all of the splice variants, probably.

Various minor fixes/adjustments.

TODO still need to adjust the functional tests to work with this!