Made mappability functions always ask for a genome seq file.
Previously there was an option to provide None, which was supposed to
use DEFAULT_ALL_GENOME_FILE which seems to have been removed from the
file earlier!  And anyway defaults are dangerous, so I modified the
functions to always require an explicit filename.

Unit-tests pass.  Also used it on real data, works.