Adjusted to multiple splice variants, + general refactoring/docstrings.
Renamed _next_gene to _next_gene_for_mapp and expanded docstring to make
it clear what's actually going on!  Also renamed curr_pos to
curr_mapp_pos for the same reason.

Added dealing with multiple-splice-variant genes to _next_gene_for_mapp,
which is somewhat complicated - we want to count them for total gene
length, but not for binned gene length if it's supposed to exclude UTRs,
and not for feature mappable length... See comments/docstring for
implementation.

In gene_mappability:
 - added dealing with the MULTIPLE_SPLICE_VARIANTS feature in
   feature_total_mappable_lengths
 - added/removed some TODO/etc comments

Updated unit-tests (ONLY to add a MULTIPLE_SPLICE_VARIANTS:0 to the
results); they pass.