Mutants with only non-conf reads have max-conf-dist 0; code cleanup.
1) Changed the max_confirmed_distance behavior:
Previously it was always NaN if there were no confirming reads, now it's
NaN if there are no uniquely aligned Carette genome-side reads
(confirming or not), and 0 if there are no confirming but some
non-confirming. This makes more sense, and makes it easier to
process/categorize the mutants later. (The change was prompted by
analysis problems on real data.)

2) Cleaned up dealing with unaligned/multi-aligned Carette genome-side
reads: Added constants for unaligned/multi-aligned positions instead of
using strings.  Changed get_Carette_pos_from_flanking_region_pos to
check that the position is one of those or an Insertion_position.
Changed the rest of the code to use the constants.

Updated unit-tests to use the new setup; added some to test 1 properly.
All pass.