Implemented dealing with distant cassette-side seqs in one mutant.
If the distant position has a lot fewer reads than the main position,
just remove it (readcount ratio is customizable).
If the read counts are similar, just remove the mutant entirely, and
print a warning (or not, if "quiet" option is on).

Updated unit-tests to use new method, i.e. to check the return value of
decide_and_check_position rather than whether it raises an exception or
not.

Also a minor bugfix in _parse_3files_parallel error message - file2_sam
and file3_sam instead of *_fasta.

All tests pass.