Added field_splitter option to parse_gene_annotation_file (space, tab, any, etc)
One of the extra files I wanted to use was separated with arbitrary
numbers of spaces instead of with tabs, but on the other hand I didn't
want to just switch from splitting on tabs to splitting on any
whitespace because that skips empty fields, so I had to implement it as
an option that can be different for each file.  Added the option to
parse_gene_annotation_file, added it to
DEFAULT_GENE_ANNOTATION_FILES_v5p5, and changed get_all_gene_annotation
to expect it in its gene_annotation_files.

Also made the mismatched-field-numbers error message more informative -
that's how I figured out what was going wrong with my file.

Tested on real data, works.