Symmetrical thresholds; print counts of 2:0, 1:0 etc mutants.
Added printed counts of mutants with 2:0, 1:0, 1:1 etc wt:phenotype
allele ratios.  In order to get those simplified counts when there are
more than 2 bins, converted some bins to wt and some to phenotype, based
on either a number of middle bins, or a fraction of mutants that should
be wt.

Changed the way phenotype thresholds work to make the symmetrical by
default (user inputs one side and the other side is auto-generated),
with an asymmetrical option.  Better to assume symmetry to avoid
p-hacking.