Added scatterplots and two-sample option; many other changes.
Fixed phenotype thresholds to be allowed to be in either direction, and
everything else to work with that.

Added scatterplots of raw and normalized readcounts with cutoffs and
hits.  Added output file.  In two-sample cases, added gene FDR
scatterplot between samples.

Fixed enrichment checks to make sense.

Changed outfiles to folder.